Fixing race condition issue while updating the partition info causing KafkaException of “There is no  metadata  for topic”
In multi-threaded use case, thread X can conclude that the partition
info is already been updated (exists and not expired) while thread Y is
in the middle of updating within the UpdateInfo method.

In this scenario, thread Y updated topicPartitionInfo and
topicPartitionInfoLastUpdateTime but have not yet updated
topicPartitionInfoList for the topic. Thread X already determined that
the update already current, will fail at GetBrokerPartitionInfo while
validating topicPartitionInfoList and throw the KafkaException "There is
no  metadata  for topic”